image_description_app/db_maintenance/requirements.txt
alembic
SQLAlchemy
psycopg2-binary
python-dotenv